# contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vis Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/assistant-turns": {
      "post": {
        "tags": [
          "assistant turns"
        ],
        "summary": "Create Assistant Turn",
        "operationId": "create_assistant_turn_api_v1_assistant_turns_post",
        "parameters": [
          {
            "name": "prefer",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prefer"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantTurnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AssistantTurnResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AssistantTurnAccepted"
                    }
                  ],
                  "title": "Response Create Assistant Turn Api V1 Assistant Turns Post"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The requested operation is not valid for the current run state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The request does not match the API contract.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The configured language-model provider failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "The language-model provider is not configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "202": {
            "description": "Async planner request accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantTurnAccepted"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant-turns/{turn_id}/trace": {
      "get": {
        "tags": [
          "assistant turns"
        ],
        "summary": "Get Assistant Turn Trace",
        "operationId": "get_assistant_turn_trace_api_v1_assistant_turns__turn_id__trace_get",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Turn Id"
            }
          },
          {
            "name": "X-Developer-Trace-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Developer-Trace-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeveloperTraceResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Developer trace access was denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant-turns/{turn_id}": {
      "get": {
        "tags": [
          "assistant turns"
        ],
        "summary": "Get Assistant State",
        "operationId": "get_assistant_state_api_v1_assistant_turns__turn_id__get",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Turn Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantTurnSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant-turns/{turn_id}/events": {
      "get": {
        "tags": [
          "assistant turns"
        ],
        "summary": "Assistant Events",
        "operationId": "assistant_events_api_v1_assistant_turns__turn_id__events_get",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Turn Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant-turns/{turn_id}/answer": {
      "post": {
        "tags": [
          "assistant turns"
        ],
        "summary": "Answer Planner",
        "operationId": "answer_planner_api_v1_assistant_turns__turn_id__answer_post",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Turn Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlannerAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantTurnAccepted"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The requested operation is not valid for the current run state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The request does not match the API contract.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant-turns/{turn_id}/cancel": {
      "post": {
        "tags": [
          "assistant turns"
        ],
        "summary": "Cancel Assistant",
        "operationId": "cancel_assistant_api_v1_assistant_turns__turn_id__cancel_post",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Turn Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantTurnSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentActivity": {
        "properties": {
          "sequence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence"
          },
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "agent",
              "tool",
              "routing",
              "question",
              "render"
            ],
            "title": "Kind"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "waiting",
              "blocked",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "tool_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Name"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sequence",
          "step_id",
          "kind",
          "actor",
          "label",
          "status",
          "occurred_at"
        ],
        "title": "AgentActivity"
      },
...
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "recoverable": {
            "type": "boolean",
            "title": "Recoverable",
            "default": false
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError"
      },
      "ApiErrorEnvelope": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelope"
      },
...
      "AssistantTurnAccepted": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "turn_id": {
            "type": "string",
            "title": "Turn Id"
          },
          "status": {
            "type": "string",
            "const": "running",
            "title": "Status",
            "default": "running"
          },
          "links": {
            "$ref": "#/components/schemas/AssistantTurnLinks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turn_id",
          "links"
        ],
        "title": "AssistantTurnAccepted"
      },
      "AssistantTurnLinks": {
        "properties": {
          "trace": {
            "type": "string",
            "title": "Trace"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "events": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events"
          },
          "cancel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trace"
        ],
        "title": "AssistantTurnLinks"
      },
      "AssistantTurnOutcome": {
        "type": "string",
        "enum": [
          "message",
          "plot_run",
          "question"
        ],
        "title": "AssistantTurnOutcome"
      },
      "AssistantTurnRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "request": {
            "$ref": "#/components/schemas/PlotRequest"
          },
          "data_scope": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AutoDataScope"
              },
              {
                "$ref": "#/components/schemas/SelectedDataScope"
              },
              {
                "$ref": "#/components/schemas/DemoDataScope"
              }
            ],
            "title": "Data Scope",
            "discriminator": {
              "propertyName": "mode",
              "mapping": {
                "auto": "#/components/schemas/AutoDataScope",
                "demo": "#/components/schemas/DemoDataScope",
                "selected": "#/components/schemas/SelectedDataScope"
              }
            }
          },
          "result_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Result Ids"
          },
          "base_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Version Id"
          },
          "parameter_changes": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "object",
            "maxProperties": 100,
            "title": "Parameter Changes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "request",
          "data_scope"
        ],
        "title": "AssistantTurnRequest"
      },
      "AssistantTurnResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "turn_id": {
            "type": "string",
            "title": "Turn Id"
          },
          "outcome": {
            "$ref": "#/components/schemas/AssistantTurnOutcome"
          },
          "intent": {
            "$ref": "#/components/schemas/IntentDecision"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "plot_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlotRunAccepted"
              },
              {
                "type": "null"
              }
            ]
          },
          "links": {
            "$ref": "#/components/schemas/AssistantTurnLinks"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlannerQuestions"
              },
              {
                "type": "null"
              }
            ]
          },
          "activity": {
            "items": {
              "$ref": "#/components/schemas/AgentActivity"
            },
            "type": "array",
            "title": "Activity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turn_id",
          "outcome",
          "intent",
          "links",
          "created_at"
        ],
        "title": "AssistantTurnResponse"
      },
      "AssistantTurnSnapshot": {
        "properties": {
          "reference_images": {
            "items": {
              "$ref": "#/components/schemas/ReferenceImage"
            },
            "type": "array",
            "title": "Reference Images"
          },
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "turn_id": {
            "type": "string",
            "title": "Turn Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "request_text": {
            "type": "string",
            "title": "Request Text"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "awaiting_input",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "activity": {
            "items": {
              "$ref": "#/components/schemas/AgentActivity"
            },
            "type": "array",
            "title": "Activity"
          },
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlannerQuestions"
              },
              {
                "type": "null"
              }
            ]
          },
          "response": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssistantTurnResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssistantFailure"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turn_id",
          "project_id",
          "request_text",
          "revision",
          "status"
        ],
        "title": "AssistantTurnSnapshot"
      },
      "AutoDataScope": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "auto",
            "title": "Mode",
            "default": "auto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AutoDataScope"
      },
...
      "DemoDataScope": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "demo",
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "DemoDataScope",
        "description": "Explicitly select illustrative data instead of project data."
      },
...
      "DeveloperTraceResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "turn_id": {
            "type": "string",
            "title": "Turn Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/DeveloperTraceEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "reasoning_content_exposed": {
            "type": "boolean",
            "const": false,
            "title": "Reasoning Content Exposed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turn_id",
          "entries"
        ],
        "title": "DeveloperTraceResponse"
      },
...
      "GenerationMode": {
        "type": "string",
        "enum": [
          "auto",
          "skill",
          "raw_code"
        ],
        "title": "GenerationMode"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlannerAnswerRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "interaction_id": {
            "type": "string",
            "title": "Interaction Id"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/ClarificationAnswer"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Answers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "interaction_id",
          "answers"
        ],
        "title": "PlannerAnswerRequest"
      },
      "PlannerQuestions": {
        "properties": {
          "interaction_id": {
            "type": "string",
            "title": "Interaction Id"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ClarificationQuestion"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Questions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "questions"
        ],
        "title": "PlannerQuestions"
      },
...
      "PlotRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "title": "Text",
            "default": ""
          },
          "reference_image_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Image Ids"
          },
          "generation_mode": {
            "$ref": "#/components/schemas/GenerationMode",
            "default": "auto"
          },
          "skill_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Id"
          },
          "gallery_mode": {
            "$ref": "#/components/schemas/GalleryMode",
            "default": "off"
          },
          "gallery_reference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gallery Reference Id"
          },
          "controls_mode": {
            "$ref": "#/components/schemas/ControlsMode",
            "default": "hybrid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PlotRequest"
      },
...
      "PlotRunAccepted": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "links": {
            "$ref": "#/components/schemas/RunLinks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "status",
          "stage",
          "links"
        ],
        "title": "PlotRunAccepted"
      },
...
      "SelectedDataScope": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "selected",
            "title": "Mode"
          },
          "bundle_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Bundle Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "bundle_ids"
        ],
        "title": "SelectedDataScope"
      },
...
    }
  }
}